iac/pyproject.toml
# ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".

[tool.poetry]
name = "iac"
version = "0.0.0"
description = "Infrastructure as Code for the RAG system"
authors = [ "Jacob Petterle <jacobpetterle@tai-tutor.team>" ]
readme = "README.md"

  [tool.poetry.dependencies]
  aws-cdk-lib = "^2.0.0"
  "aws-cdk.aws-lambda-python-alpha" = "^2.153.0a0"
  pinecone-db-construct = "^0.0.10"
  python = "^3.9"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.3"

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"
